# api/openapi.yaml
openapi: 3.0.3
info:
  title: Meetopoly API
  version: 0.5.0
  description: |
    Meetopoly HTTP API contract.
    Source of truth for mobile codegen (orval → meetopoly-mobile/api/).
  contact:
    name: Meetopoly
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /auth/signup/start:
    post:
      operationId: signupStart
      summary: Start signup by email
      description: |
        Accepts an email and sends a 6-digit verification code via SMTP.
        Always returns 204 when the request is well-formed (anti-enumeration).
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupStartRequest"
      responses:
        "204":
          description: Verification email sent if the address is eligible
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Mail or persistence failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup/verify:
    post:
      operationId: signupVerify
      summary: Verify email with 6-digit code
      description: On success returns a short-lived signup token for password + profile steps.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupVerifyRequest"
      responses:
        "200":
          description: Email verified; signup token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupVerifyResponse"
        "400":
          description: Invalid request or code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup/password:
    post:
      operationId: signupSetPassword
      summary: Set password during signup
      description: Requires Bearer signup token from `/auth/signup/verify`.
      tags:
        - Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupPasswordRequest"
      responses:
        "204":
          description: Password set
        "400":
          description: Weak or invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid signup token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup/profile:
    post:
      operationId: signupCompleteProfile
      summary: Finish profile and start session
      description: |
        Requires Bearer signup token. Sets username + country and returns an opaque session token.
      tags:
        - Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupProfileRequest"
      responses:
        "200":
          description: Profile complete; session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "400":
          description: Invalid username or country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid signup token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup/status:
    get:
      operationId: signupStatus
      summary: Signup progress for a signup Bearer token
      description: |
        Used to resume onboarding. Only `passwordSet: true` is durable progress;
        OTP-only accounts should discard the draft and stay on login.
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current signup progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupStatusResponse"
        "401":
          description: Missing or invalid signup token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      summary: Login with email and password
      description: |
        Requires verified email and a password. When the profile is incomplete,
        returns `needsProfile: true` plus a reissued signup token instead of a session.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated, or resume profile step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials or incomplete signup (no password yet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke current session
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session revoked (idempotent if already gone)
        "401":
          description: Missing or invalid session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password-reset/start:
    post:
      operationId: passwordResetStart
      summary: Start password reset by email
      description: |
        Sends a 6-digit code only when the email belongs to a fully registered user.
        Response `sent` is false when no mail was sent (unknown or incomplete account).
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetStartRequest"
      responses:
        "200":
          description: Whether a reset email was sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetStartResponse"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Mail or persistence failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password-reset/verify:
    post:
      operationId: passwordResetVerify
      summary: Verify password-reset code
      description: On success returns a short-lived reset token for setting a new password.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetVerifyRequest"
      responses:
        "200":
          description: Code verified; reset token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetVerifyResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password-reset/confirm:
    post:
      operationId: passwordResetConfirm
      summary: Set a new password after reset verify
      description: |
        Requires Bearer reset token. Updates the password and revokes all existing login sessions.
      tags:
        - Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequest"
      responses:
        "204":
          description: Password updated; sessions revoked
        "400":
          description: Weak or invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid reset token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me:
    get:
      operationId: getMe
      summary: Current user profile
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Missing or invalid session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /worlds:
    get:
      operationId: listWorlds
      summary: List world packs
      description: Distinct `worldId` values with location counts from the `locations` collection.
      tags:
        - Worlds
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Worlds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldsResponse"
        "401":
          description: Missing or invalid session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tables/join:
    post:
      operationId: joinTable
      summary: Join a World matchmaking lobby
      description: |
        Finds an open lobby for `worldId` (or creates one), seats the caller,
        and returns the table snapshot. Connect to `GET /ws/tables/{tableId}?token=`
        for live seat/ready updates. Messages: client `{type:ready|leave|ping}`,
        server `{type:state|started|pong|error}`.
      tags:
        - Tables
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinTableRequest"
      responses:
        "200":
          description: Seated in lobby
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "400":
          description: Invalid world
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Table full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tables/{tableId}:
    get:
      operationId: getTable
      summary: Get table snapshot
      tags:
        - Tables
      security:
        - bearerAuth: []
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tables/{tableId}/leave:
    post:
      operationId: leaveTable
      summary: Leave lobby (frees seat immediately)
      tags:
        - Tables
      security:
        - bearerAuth: []
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated table after leave
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /games/{gameId}:
    get:
      operationId: getGame
      summary: Get M1 game snapshot
      description: |
        Authoritative game state (players, MeetCoin balances, pin boardIndex, whose turn).
        Created when a lobby reaches all-Ready (Phase 6.0).
      tags:
        - Games
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations:
    get:
      operationId: listLocations
      summary: List locations for a world
      description: Returns board spaces for `worldId`, sorted by `boardIndex`.
      tags:
        - Locations
      security:
        - bearerAuth: []
      parameters:
        - name: worldId
          in: query
          required: true
          schema:
            type: string
            example: africa-1
          description: World pack id (e.g. africa-1)
      responses:
        "200":
          description: Locations for the world
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationsResponse"
        "400":
          description: Missing worldId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/by-slug:
    get:
      operationId: getLocationBySlug
      summary: Get location by world and slug
      tags:
        - Locations
      security:
        - bearerAuth: []
      parameters:
        - name: worldId
          in: query
          required: true
          schema:
            type: string
            example: africa-1
        - name: slug
          in: query
          required: true
          schema:
            type: string
            example: accra
      responses:
        "200":
          description: Location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "400":
          description: Missing worldId or slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/{locationId}:
    get:
      operationId: getLocationById
      summary: Get location by Mongo id
      tags:
        - Locations
      security:
        - bearerAuth: []
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId hex string
      responses:
        "200":
          description: Location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "401":
          description: Missing or invalid session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque session token (or short-lived signup token) stored in Redis
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable detail
    SignupStartRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: Account email address
    SignupVerifyRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: "^[0-9]{6}$"
          description: 6-digit verification code from email
    SignupVerifyResponse:
      type: object
      additionalProperties: false
      required:
        - signupToken
      properties:
        signupToken:
          type: string
          description: Short-lived Bearer token for password + profile steps
    SignupPasswordRequest:
      type: object
      additionalProperties: false
      required:
        - password
      properties:
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
    SignupProfileRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - country
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          pattern: "^[A-Za-z0-9_]{3,20}$"
          description: Unique username (letters, digits, underscore)
        country:
          type: string
          pattern: "^[A-Z]{2}$"
          description: ISO 3166-1 alpha-2 country code (uppercase)
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      additionalProperties: false
      required:
        - needsProfile
        - user
      properties:
        needsProfile:
          type: boolean
          description: True when password is set but username/country are missing
        token:
          type: string
          nullable: true
          description: Opaque session token when needsProfile is false
        signupToken:
          type: string
          nullable: true
          description: Reissued signup Bearer token when needsProfile is true
        user:
          $ref: "#/components/schemas/UserProfile"
    SignupStatusResponse:
      type: object
      additionalProperties: false
      required:
        - email
        - passwordSet
        - profileComplete
      properties:
        email:
          type: string
          format: email
        passwordSet:
          type: boolean
        profileComplete:
          type: boolean
    PasswordResetStartRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
...
    AuthSessionResponse:
      type: object
      additionalProperties: false
      required:
        - token
        - user
      properties:
        token:
          type: string
          description: Opaque session token (store in expo-secure-store)
        user:
          $ref: "#/components/schemas/UserProfile"
...
    JoinTableRequest:
      type: object
      additionalProperties: false
      required:
        - worldId
      properties:
        worldId:
          type: string
          example: africa-1
...